Improvement

+++ src/AppBundle/Controller/TaskController.php
@@ -30,7 +30,7 @@ class TaskController extends Controller
 
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
-
+            $task->setUser($this->getUser());
             $em->persist($task);
             $em->flush();
 
@@ -83,12 +83,15 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
-        $em = $this->getDoctrine()->getManager();
-        $em->remove($task);
-        $em->flush();
-
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
-
+        $flash = ['type' => 'error', 'message' => "Vous n'avez pas les autorisations requises"];
+        if($task->getUser() === $this->getUser()  || $task->hasAnonUser() && $this->getUser()->getRoles()[0] === 'ROLE_ADMIN')
+        {
+            $flash = ['type' => 'success', 'message' => "La tâche a bien été supprimée."];
+            $em = $this->getDoctrine()->getManager();
+            $em->remove($task);
+            $em->flush();
+        }
+        $this->addFlash($flash['type'], $flash['message']);
         return $this->redirectToRoute('task_list');
     }
 }
